src/model4/semanticmodel_final0210/requirements.txt
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
